홈 편집 오류 수정
