+++ src/components/template/home/HomeTemplate.tsx
@@ -3,32 +3,23 @@ import PlaylistBox from '@/components/pages/home/playlist-box';
 import { useState } from 'react';
 
 interface HomeTemplateProps {
-  isPlaylistSelected: boolean;
   whenPlaylistWillPlay: string;
   isTimeBadgeVisible: boolean;
-  isEditSelected: boolean;
   playlistImageSources: string[];
-  isAlarmSelected: boolean;
   isPlaylistAvailable: boolean;
   isEditMode: boolean;
   onClickIsEditMode : () => void;
-  onClickAlarmButton : () => void;
-  onClickPlaylistSelectButton: () => void;
   TestPlayListData : any;
 }
 
+
 export const HomeTemplate = ({
-  isPlaylistSelected,
   whenPlaylistWillPlay,
   isTimeBadgeVisible,
-  isEditSelected,
   playlistImageSources,
-  isAlarmSelected,
   isPlaylistAvailable,
   isEditMode,
   onClickIsEditMode,
-  onClickAlarmButton,
-  onClickPlaylistSelectButton,
   TestPlayListData
 }: HomeTemplateProps) => {
   
@@ -42,44 +33,29 @@ export const HomeTemplate = ({
       { !isEditMode?
       <style.Button
        onClick = {onClickIsEditMode}
+       isEditMode={isEditMode}
       >편집</style.Button> : 
       <style.Button
        onClick = {onClickIsEditMode}
+       isEditMode={isEditMode}
       >취소</style.Button> }
       </style.EditDiv>
       <PlaylistBox
-        isPlaylistSelected={isPlaylistSelected}
         whenPlaylistWillPlay={whenPlaylistWillPlay}
         isTimeBadgeVisible={isTimeBadgeVisible}
-        isEditSelected={isEditSelected}
         playlistImageSources={playlistImageSources}
-        isAlarmSelected={isAlarmSelected}
         isEditMode ={isEditMode}
         TestPlayListData={TestPlayListData}
-        onClickAlarmButton ={onClickAlarmButton}
-        onClickPlaylistSelectButton ={onClickPlaylistSelectButton}
-      /> 
-      {/* 테스트용 */}
-      <PlaylistBox
-        isPlaylistSelected={isPlaylistSelected}
-        whenPlaylistWillPlay={whenPlaylistWillPlay}
-        isTimeBadgeVisible={isTimeBadgeVisible}
-        isEditSelected={isEditSelected}
-        playlistImageSources={playlistImageSources}
-        isAlarmSelected={isAlarmSelected}
-        isEditMode ={isEditMode}
-        TestPlayListData={TestPlayListData}
-        onClickAlarmButton ={onClickAlarmButton}
-        onClickPlaylistSelectButton ={onClickPlaylistSelectButton}
-      /> 
+        /> 
+     
       </style.ShowPlaylist>
       : 
-       <style.SuggestBox>
        <style.SuggestAddPlaylist>콤마 플레이리스트가 없어요.</style.SuggestAddPlaylist>
-       <style.AddPlaylistButton></style.AddPlaylistButton>
-       </style.SuggestBox>
       }
-
+      { !isEditMode?
+      <style.AddPlaylistButton/> : <style.DeletePlaylistButton/>
+      }
+      
     </style.Container>
   );
 };
